Отладка тестов
